REM-812: Change ws architecture structure
