+++ config/remme-rpc-api.toml
@@ -6,7 +6,7 @@ port = 8080
 exports_folder = "./default_export"
 
 # List of available modules for rpc
-available_modules = "pkc,transaction,block_info,account,network,atomic_swap,state"
+available_modules = "pkc,transaction,block_info,account,network,atomic_swap,state,event"
 
 [remme.rpc_api.cors]
 # The origin, or list of origins to allow requests from.
