+++ pyproject.toml
@@ -23,7 +23,6 @@ aiohttp = "3.2.1"
 aiohttp-jinja2 = "1.0.0"
 aiohttp-cors = "0.7.0"
 aiohttp-json-rpc = "0.11.2"
-ujson = "1.35"
 toml = "0.9.4"
 pysha3 = "1.0.2"
 aiozmq = "0.7.1"
