+++ remme/clients/atomic_swap.py
@@ -144,10 +144,14 @@ class AtomicSwapClient(BasicClient):
 
     async def swap_get(self, swap_id):
         atomic_swap_info = AtomicSwapInfo()
-        atomic_swap_info.ParseFromString(await self.get_value(self.make_address_from_data(swap_id)))
+        swap_address = self.make_address_from_data(swap_id)
+        raw_swap = await self.get_value(swap_address)
+        atomic_swap_info.ParseFromString(raw_swap)
         return atomic_swap_info
 
     async def get_pub_key_encryption(self):
         setting = Setting()
-        setting.ParseFromString(await self.get_value(_make_settings_key(SETTINGS_PUB_KEY_ENCRYPTION)))
+        pub_key_enc = _make_settings_key(SETTINGS_PUB_KEY_ENCRYPTION)
+        raw_pub_key = await self.get_value(pub_key_enc)
+        setting.ParseFromString(raw_pub_key)
         return setting.entries[0].value
